requirements.txt
py-cord
python-escpos>=3.0.0
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
python-multipart>=0.0.5
Pillow>=8.0.0
requests>=2.25.0
